+++ indigo-frontend/src/app/pages/project/project-info/project-info.component.html
@@ -43,7 +43,7 @@
         @if (project.description || project.details) {
           <div>
             <h3 class="text-neutral-800 mb-2 font-semibold">Project Details</h3>
-            <p class="text-neutral-1000">{{ project.description }}</p>
+            <p class="text-neutral-1000" [innerHTML]="project.description"></p>
             <p class="text-neutral-1000 mt-4">{{ project.details }}</p>
           </div>
         }
